refactor: service 나뉘었던거 통합
