+++ src/main/java/chess/controller/api/LobbyController.java
@@ -3,7 +3,7 @@ package chess.controller.api;
 import chess.dto.ChessGameDto;
 import chess.dto.ChessGamesDto;
 import chess.dto.TitleDto;
-import chess.service.LobbyService;
+import chess.service.GameService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,22 +16,22 @@ import java.util.stream.Collectors;
 @RestController
 public class LobbyController {
 
-    private final LobbyService lobbyService;
+    private final GameService gameService;
 
-    public LobbyController(LobbyService lobbyService) {
-        this.lobbyService = lobbyService;
+    public LobbyController(GameService gameService) {
+        this.gameService = gameService;
     }
 
     @PostMapping("/game")
     public ResponseEntity<Long> newGame(@RequestBody @Valid TitleDto titleDto) {
-        lobbyService.verifyDuplicateTitleInGames(titleDto.getTitle());
-        return ResponseEntity.ok(lobbyService.newGame(titleDto.getTitle()));
+        gameService.verifyDuplicateTitleInGames(titleDto.getTitle());
+        return ResponseEntity.ok(gameService.newGame(titleDto.getTitle()));
     }
 
     @GetMapping("/games")
     public ResponseEntity<ChessGamesDto> findAllRooms() {
         return ResponseEntity.ok(new ChessGamesDto(
-                lobbyService.findAllGames()
+                gameService.findAllGames()
                         .stream()
                         .map(ChessGameDto::new)
                         .collect(Collectors.toList())
